project/backend/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.4.0
PyMySQL==1.1.1
celery==5.3.4
redis==5.0.8
boto3==1.35.10
python-magic==0.4.27
structlog==23.2.0
sentry-sdk==2.13.0
python-dotenv==1.0.1
black==24.8.0
ruff==0.6.2
pytest==8.3.2
pytest-django==4.8.0
factory-boy==3.3.1